Add External IDs to goodies
