+++ kifi-backend/modules/shoebox/app/com/keepit/admin/AdminGoodiesController.scala
@@ -4,10 +4,16 @@ import java.util.regex.{ PatternSyntaxException, Pattern }
 
 import com.google.inject.Inject
 import com.keepit.common.controller.{ UserActionsHelper, AdminUserActions }
+import com.keepit.common.crypto.{ PublicIdRegistry, PublicIdConfiguration }
+import com.keepit.common.db.Id
+import com.keepit.model.{ Organization, Library }
 import play.api.libs.json.Json
 
+import scala.util.Try
+
 class AdminGoodiesController @Inject() (
-    override val userActionsHelper: UserActionsHelper) extends AdminUserActions {
+    override val userActionsHelper: UserActionsHelper,
+    implicit val publicIdConfig: PublicIdConfiguration) extends AdminUserActions {
 
   def testRegex = AdminUserPage { implicit request =>
     Ok(views.html.admin.roverTestRegex())
@@ -28,6 +34,42 @@ class AdminGoodiesController @Inject() (
     } catch {
       case e: PatternSyntaxException =>
         BadRequest(Json.toJson(e.toString))
+    }
+  }
+
+  val warmUp = {
+    // This just forces registration. Probably not necessary if they're actually being used, but hey.
+    Library.publicId(Id(1))
+    Organization.publicId(Id(1))
+  }
+
+  def listAllExternalIds = AdminUserPage { implicit request =>
+    Ok(registry.mkString("\n"))
+  }
+
+  def externalIdById(name: String, id: Long) = AdminUserPage { implicit request =>
+    val pubIds = PublicIdRegistry.registry.filter(_._1.toLowerCase.contains(name.toLowerCase)).map {
+      case (clazz, accessor) =>
+        clazz + " " + accessor.toPubId(id)
+    }.mkString("\n")
+    Ok(pubIds)
+  }
+
+  def externalIdByPublicId(name: String, publicId: String) = AdminUserPage { implicit request =>
+    val id = PublicIdRegistry.registry.filter(_._1.toLowerCase.contains(name.toLowerCase)).map {
+      case (clazz, accessor) =>
+        clazz + " " + Try(accessor.toId(publicId)).toOption.getOrElse("(invalid)")
+    }.mkString("\n")
+    Ok(id)
+  }
+
+  private def registry: Seq[(String, String, Long)] = {
+    PublicIdRegistry.registry.map {
+      case (companion, accessor) =>
+        val a = accessor.toPubId(1)
+        val b = accessor.toId(a)
+
+        (companion, a, b)
     }
   }
