+++ kifi-backend/modules/shoebox/app/views/admin/admin.scala.html
@@ -162,6 +162,7 @@
                 <a class="dropdown-toggle" data-toggle="dropdown">Goodies</a>
                 <ul class="dropdown-menu" role="menu">
                     <li><a href="@com.keepit.controllers.admin.routes.AdminGoodiesController.testRegex()">Test Regex</a></li>
+                    <li><a href="@com.keepit.controllers.admin.routes.AdminGoodiesController.listAllExternalIds()">External IDs</a></li>
                 </ul>
             </li>
         </ul>
